feat: data structure and endpoint definition changed
Now there is only one Chat and Message model as it's much easier when
you don't have 2 models, which to essintially the same thing in the end.
Now the routes are assigned to the router using the
APIRouter.add_api_route method. This enables us to avoid creating a
router in each file, which reduces boilderplate and look prettier when
you have all the routes defined in one file and the resolvers in the
endpoints directory. Abstracted some repetitive logic with check if the
User is in the Chat in some of the resolvers into a separate
dependencies which utilize fastapi's dependency injection api.